\name{coral}
\alias{coral}
\title{Coral-Parrotfish model}
\usage{
  coral(x, h, p)
}
\arguments{
  \item{x}{vector of population levels: macroalgae, coral,
  parrotfish}

  \item{h}{harvesting effort on parrotfish}

  \item{p}{c(a, g, T, gamma, r, d, s, K) 1 2 3 4 5 6 7 8}
}
\description{
  Coral-Parrotfish model
}

