% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SDP_uniform.R
\name{int_f}
\alias{int_f}
\title{Integrate against uniform distributions of measurement and implementation uncertainty}
\usage{
int_f(f, x, q, sigma_m, sigma_i, pars)
}
\description{
Integrate against uniform distributions of measurement and implementation uncertainty
}

