% Generated by roxygen2 (4.0.1): do not edit by hand
\name{May}
\alias{May}
\title{Harvesting model with a bifurcation from May (1977)
increasing parameter "a" (p[3]) will drive the bifurcation}
\usage{
May(x, h, p)
}
\arguments{
\item{x}{the current population level}

\item{h}{total harvest level}

\item{p}{vector of parameters c(r, K, a, H, Q)}
}
\value{
the population level in the next timestep
}
\description{
Harvesting model with a bifurcation from May (1977)
increasing parameter "a" (p[3]) will drive the bifurcation
}
\details{
try pars <- c(r = .75, k = 10, a=1, H=1, Q = 3)
See the position of the bifurcation at around a = 1.9:
curve(.75*(1-x/10), 0, 10)
curve(1.9*x^2/(x^3+1), 0, 10, add=T, col="red")
goes to nonzero alternate stable state.
}

