% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Myer_harvest}
\alias{Myer_harvest}
\title{Discrete-time model with an allee effect for alpha > 1 with harvest control}
\usage{
Myer_harvest(x, h, p)
}
\arguments{
\item{x}{the current population level}

\item{h}{total harvest level}

\item{p}{vector of parameters c(r, alpha, K)}
}
\value{
the population level in the next timestep
}
\description{
Discrete-time model with an allee effect for alpha > 1 with harvest control
}
\details{
A Beverton-Holt style model with Allee effect.
  Unharvested carrying capacity is:
  K <- p[1] * p[3] / 2 + sqrt( (p[1] * p[3]) ^ 2 - 4 * p[3] ) / 2
  The (unharvested) allee theshold is given by:
  x = p[1] * p[3] / 2 - sqrt( (p[1] * p[3]) ^ 2 - 4 * p[3] ) / 2
}

