% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Myers}
\alias{Myers}
\title{Discrete-time model with an allee effect for alpha > 1}
\usage{
Myers(x, h, p)
}
\arguments{
\item{x}{the current population level}

\item{h}{harvest effort}

\item{p}{vector of parameters c(r, alpha, K)}
}
\value{
the population level in the next timestep
}
\description{
Discrete-time model with an allee effect for alpha > 1
}
\details{
Unharvested carrying capacity is:
  K <- p[1] * p[3] / 2 + sqrt( (p[1] * p[3]) ^ 2 - 4 * p[3] ) / 2
  The (unharvested) allee theshold is given by:
  x = p[1] * p[3] / 2 - sqrt( (p[1] * p[3]) ^ 2 - 4 * p[3] ) / 2
  Bifurcation pt is h = (p[1]*sqrt(p[3])-2)/2
  Try with pars = c(1,2,6), h=.01

  Consider updating to be a function of x-h, instead?
}

