% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SDP_by_simulation.R
\name{SDP_by_simulation}
\alias{SDP_by_simulation}
\title{Determine the transtion matrix using stochastic simulation}
\usage{
SDP_by_simulation(f, p, x_grid, h_grid, z_g, z_m, z_i, reps = 999)
}
\arguments{
\item{f}{the growth function of the escapement population (x-h)
should be a function of f(t, y, p), with parameters p}

\item{p}{the parameters of the growth function}

\item{x_grid}{the discrete values allowed for the population size, x}

\item{h_grid}{the discrete values of harvest levels to optimize over}

\item{z_g}{a function determing the probability distrubtion for the
stochastic population growth process (draws a random variable z_g)}

\item{z_m}{a function determining the porbability distribution for
measurement uncertainty in the assessment of stock size (random variable)}

\item{z_i}{function for implementation uncertainty in quotas}
}
\value{
the transition matrix at each value of h in the grid.
}
\description{
Determine the transtion matrix using stochastic simulation
}

