% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/value_iteration.R
\name{value_iteration}
\alias{value_iteration}
\title{Identify the dynamic optimum using backward iteration (dynamic programming)}
\usage{
value_iteration(SDP_Mat, x_grid, h_grid, OptTime = 100, xT, profit, delta,
  epsilon = 1e-04)
}
\arguments{
\item{SDP_Mat}{the stochastic transition matrix at each h value}

\item{x_grid}{the discrete values allowed for the population size, x}

\item{h_grid}{the discrete values of harvest levels to optimize over}

\item{OptTime}{the stopping time}

\item{xT}{the boundary condition population size at OptTime}

\item{profit}{the profit function
(i.e. enforces the boundary condition)}

\item{delta}{the discounting rate (1-delta)}

\item{epsilon}{value iteration tolerance}
}
\value{
list containing the matrices D and V.  D is an x_grid by OptTime
 matrix with the indices of h_grid giving the optimal h at each value x
 as the columns, with a column for each time.
 V is a matrix of x_grid by x_grid, which is used to store the value
 function at each point along the grid at each point in time.
 The returned V gives the value matrix at the first (last) time.
}
\description{
Identify the dynamic optimum using backward iteration (dynamic programming)
}

