% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_uncertainty_SDP.R
\name{parameter_uncertainty_SDP}
\alias{parameter_uncertainty_SDP}
\title{Computes the transition matrix under the case of parametric uncertainty}
\usage{
parameter_uncertainty_SDP(f, x_grid, h_grid, pardist, sigma_g_index,
  n_mc = 100)
}
\arguments{
\item{f}{the growth function of the escapement population (x-h)
should be a function of f(t, y, pardist[i,]), with parameters}

\item{x_grid}{the discrete values allowed for the population size, x}

\item{h_grid}{the discrete values of harvest levels to optimize over}

\item{pardist}{a matrix with columns as variables and rows as the
Monte Carlo samples from the posterior}

\item{sigma_g_index}{column number containing growth noise}

\item{n_mc}{number of monte carlo replicates to sample}
}
\description{
Computes the transition matrix under the case of parametric uncertainty
}

