% Generated by roxygen2 (4.0.1): do not edit by hand
\name{profit_harvest}
\alias{profit_harvest}
\title{Define a profit function, price minus cost}
\usage{
profit_harvest(price = 1, c0 = 0.1, c1 = 0)
}
\arguments{
\item{price}{market price (Note, optimal will scrap xT if price is
high enough!)}

\item{c0}{Cost, linear with effort (cost = (c0 + c1*E)*E}

\item{c1}{Cost, quadratic with effort (cost = (c0 + c1*E)*E}
}
\value{
a function computing profit at harvest intensity h_i over
possible stock values x_grid, profit(x_grid, h_i)
}
\description{
Define a profit function, price minus cost
}

