% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/population_models.R
\name{Ricker}
\alias{Ricker}
\title{Basic Ricker model}
\usage{
Ricker(x, h, p)
}
\arguments{
\item{x}{the current population level}

\item{h}{harvest level}

\item{p}{a vector of parameters c(r, K)}
}
\value{
the population level in the next timestep
}
\description{
Basic Ricker model
}

