% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/population_models.R
\name{RickerAllee}
\alias{RickerAllee}
\title{Ricker-like model with Allee effect (Allen)}
\usage{
RickerAllee(x, h, p)
}
\arguments{
\item{x}{the current population level}

\item{h}{harvest level}

\item{p}{a vector of parameters c(r, K, C)}
}
\value{
the population level in the next timestep
}
\description{
Ricker-like model with Allee effect (Allen)
}

